hook up the performance marks for runic power
