+++ src/analysis/retail/deathknight/blood/modules/runicpower/RunicPowerTracker.ts
@@ -1,10 +1,12 @@
 import SPELLS from 'common/SPELLS';
+import Spell from 'common/SPELLS/Spell';
 import TALENTS from 'common/TALENTS/deathknight';
 import RESOURCE_TYPES from 'game/RESOURCE_TYPES';
 import { Options } from 'parser/core/Analyzer';
 import { CastEvent } from 'parser/core/Events';
 import ResourceTracker from 'parser/shared/modules/resources/resourcetracker/ResourceTracker';
 import SpellUsable from 'parser/shared/modules/SpellUsable';
+import { QualitativePerformance } from 'parser/ui/QualitativePerformance';
 
 // Vamp blood CDR per 10 RP spent by # of talent points taken
 const RED_THIRST_REDUCTION_MS: Record<number, number> = { 1: 1000, 2: 2000 };
@@ -76,6 +78,58 @@ class RunicPowerTracker extends ResourceTracker {
 
   get cooldownReductionWasted() {
     return this.totalCooldownReductionWasted;
+  }
+
+  get wastePerformance(): QualitativePerformance {
+    if (this.percentWasted < 0.02) {
+      return QualitativePerformance.Perfect;
+    } else if (this.percentWasted < 0.05) {
+      return QualitativePerformance.Good;
+    } else if (this.percentWasted < 0.1) {
+      return QualitativePerformance.Ok;
+    } else {
+      return QualitativePerformance.Fail;
+    }
+  }
+
+  baseGeneratedRp(spells: Spell[]): number {
+    return spells
+      .map((spell) => {
+        const casts = this.buildersObj[spell.id]?.casts ?? 0;
+        return casts * (spell.runesCost ?? 0) * 10;
+      })
+      .reduce((a, b) => a + b, 0);
+  }
+
+  generatedRp(spells: Spell[]): number {
+    return spells
+      .map((spell) => {
+        return this.buildersObj[spell.id]?.generated ?? 0;
+      })
+      .reduce((a, b) => a + b, 0);
+  }
+
+  wastedRp(spells: Spell[]): number {
+    return spells
+      .map((spell) => {
+        return this.buildersObj[spell.id]?.wasted ?? 0;
+      })
+      .reduce((a, b) => a + b, 0);
+  }
+
+  bonusGeneratedRp(spells: Spell[]): number {
+    const base = this.baseGeneratedRp(spells);
+    const actual = this.generatedRp(spells) + this.wastedRp(spells);
+    return actual - base;
+  }
+
+  estRunesSpent(spells: Spell[]): number {
+    return spells
+      .map((spell) => {
+        const casts = this.buildersObj[spell.id]?.casts ?? 0;
+        return casts * (spell.runesCost ?? 0);
+      })
+      .reduce((a, b) => a + b, 0);
   }
 }
